requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Markdown==2.6.2
Pygments==2.0.2
PyYAML==3.11
uWSGI==2.0.11.2
